requirements.txt
aioec
ben_cogs
discord.py @ git+https://github.com/Rapptz/discord.py@rewrite
jishaku
simple_help_formatter @ git+https://github.com/bmintz/simple-dpy-helpformatter
wand
